+++ backend/package.json
@@ -42,8 +42,8 @@
     "fs-extra": "^11.3.0",
     "ioredis": "^5.6.1",
     "koa": "^2.16.1",
-    "koa-body": "^2.6.0",
-    "koa-logger": "^3.2.1",
+    "koa-body": "^6.0.1",
+    "koa-logger": "^4.0.0",
     "koa-send": "^5.0.1",
     "koa-session": "^7.0.2",
     "koa-static": "^5.0.0",
@@ -65,7 +65,7 @@
     "@types/koa": "^2.15.0",
     "@types/koa__router": "^12.0.4",
     "@types/lodash": "^4.17.20",
-    "@types/node": "^24.0.14",
+    "@types/node": "^24.0.15",
     "ava": "^6.4.1",
     "eslint": "^9.31.0",
     "eslint-plugin-import": "^2.32.0",
