+++ backend/src/app.js
@@ -3,7 +3,7 @@ require('dotenv-flow').config()
 const path = require('node:path')
 const process = require('node:process')
 const Koa = require('koa')
-const koaBody = require('koa-body')
+const { koaBody } = require('koa-body')
 const koaLogger = require('koa-logger')
 const send = require('koa-send')
 const { createSession: session } = require('koa-session')
